fix: file names, function names, makefile, dir tree, and lst libft function wrapp
