+++ .gitignore
@@ -1,0 +1,3 @@
+*.dist
+.vscode
+.DS_Store
