+++ pushswap/inc/push_swap.h
@@ -3,17 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   push_swap.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ivbatist <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: pin3dev <pinedev@outlook.com>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/17 22:47:30 by ivbatist          #+#    #+#             */
-/*   Updated: 2023/05/17 22:47:39 by ivbatist         ###   ########.fr       */
+/*   Updated: 2024/10/01 00:36:48 by pin3dev          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-#ifndef PUSH_SWAP_H
+
+# ifndef PUSH_SWAP_H
 # define PUSH_SWAP_H
 
 # include<stdlib.h>
 # include<unistd.h>
+# include "../../../42_Libft/libft/inc/libft.h"
+
 //////////////////////////////////////////////////ESTRUTURA
 typedef struct stack
 {
@@ -25,26 +28,27 @@ typedef struct stack
 	int				index;
 }			t_stack;
 //////////////////////////////////////////////////FT GERAIS
+int	pushswap(int ac, char **av);
 void		ft_error(void);
 int			ft_isdigit(int c);
 //////////////////////////////////////////////////FT GERAIS LISTA
-t_stack		*ft_lstnew(int content);
-void		ft_lstadd_back(t_stack **lst, t_stack *new);
-t_stack		*ft_lstlast(t_stack *lst);
-int			ft_lstsize(t_stack *lst);
+t_stack		*ft_stacknew(int content);
+void		ft_stackadd_back(t_stack **lst, t_stack *new);
+t_stack		*ft_stacklast(t_stack *lst);
+int			ft_stacksize(t_stack *lst);
 //////////////////////////////////////////////////FT VERIFICA MAIN
-int			ft_atoi(const char *str);
+int			_atoi(const char *str);
 const char	*ft_skip_spaces(const char *str);
 size_t		ft_strlen(const char *str);
 char		*ft_substr(char const *s, unsigned int start, size_t len);
 char		**ft_split(char const *s, char c);
-void		ft_create_node(t_stack **head, char *str);
-void		ft_freelst(t_stack **stack);
-void		ft_freesplit(char **str);
-t_stack		*ft_check_arg_and_create_stack_a(int ac, char **av);
-t_stack		*ft_args_splited(char **split);
-int			ft_check_doubles(t_stack **stack_a);
-int			ft_check_sorted(t_stack **stack);
+void		create_node(t_stack **head, char *str);
+void		free_stack(t_stack **stack);
+void		free_split(char **str);
+t_stack		*args_to_node(int ac, char **av);
+t_stack		*args_splited(char **split);
+int			ck_doubles(t_stack **stack_a);
+int			ck_sorted(t_stack **stack);
 //////////////////////////////////////////////////FT MOVIMENTO
 void		swap(t_stack **stack);
 void		rotate(t_stack **stack);
@@ -64,22 +68,22 @@ void		pb(t_stack **b, t_stack **a);
 //////////////////////////////////////////////////FT ALGORITMO
 t_stack		*ft_max(t_stack **stack);
 t_stack		*ft_min(t_stack **stack);
-t_stack		*ft_check_new_min_max(t_stack *node_a, t_stack **stack_b);
-t_stack		*ft_return_dest_place(t_stack *src, t_stack **dest);
-void		ft_find_place_n_def_total_mov(t_stack **src, t_stack **dest);
-int			ft_put_index_n_def_mov_n_orient(t_stack **src, t_stack **dest);
-void		ft_put_mov_and_orientation(int size, t_stack **stack);
-void		ft_put_total_mov(t_stack *src, t_stack *dest);
-t_stack		*ft_find_best_case(t_stack **src);
-void		ft_move_node_to_top_a(t_stack *node_a, t_stack**stack_a);
-void		ft_move_node_to_top_b(t_stack *node_b, t_stack**stack_b);
-void		ft_move_node_to_top_ab(t_stack *node_a, t_stack *node_b, \
+t_stack		*ck_new_min_max(t_stack *node_a, t_stack **stack_b);
+t_stack		*return_dest_place(t_stack *src, t_stack **dest);
+void		find_place_n_def_total_mov(t_stack **src, t_stack **dest);
+int			put_index_n_def_mov_n_orient(t_stack **src, t_stack **dest);
+void		put_mov_and_orientation(int size, t_stack **stack);
+void		put_total_mov(t_stack *src, t_stack *dest);
+t_stack		*find_best_case(t_stack **src);
+void		mv_node_to_top_a(t_stack *node_a, t_stack**stack_a);
+void		mv_node_to_top_b(t_stack *node_b, t_stack**stack_b);
+void		mv_node_to_top_ab(t_stack *node_a, t_stack *node_b, \
 t_stack**stack_a, t_stack**stack_b);
-t_stack		*ft_fit_between(t_stack *src, t_stack **dest);
-void		ft_rotate_to_finish(t_stack **stack_a, t_stack **stack_b);
-void		ft_back_to_a(t_stack **stack_a, t_stack **stack_b);
-void		ft_sort_3(t_stack **stack_a);
-void		ft_big_sort(t_stack **stack_a, t_stack **stack_b);
-void		ft_sort_cases(t_stack **stack_a);
+t_stack		*fit_between(t_stack *src, t_stack **dest);
+void		rotate_to_finish(t_stack **stack_a, t_stack **stack_b);
+void		back_to_a(t_stack **stack_a, t_stack **stack_b);
+void		sort_3(t_stack **stack_a);
+void		big_sort(t_stack **stack_a, t_stack **stack_b);
+void		sort_cases(t_stack **stack_a);
 
 #endif
