--- srcs/ft_split_functions.c
@@ -1,114 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_split_functions.c                               :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: pin3dev <pinedev@outlook.com>              +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/17 20:43:27 by ivbatist          #+#    #+#             */
-/*   Updated: 2024/08/08 15:05:37 by pin3dev          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../inc/push_swap.h"
-
-static int	count_words(char const *str, char c)
-{
-	int	words;
-	int	flag;
-
-	words = 0;
-	flag = 0;
-	while (*str)
-	{
-		if (*str != c && flag == 0)
-		{
-			flag = 1;
-			words++;
-		}
-		else if (*str == c)
-			flag = 0;
-		str++;
-	}
-	return (words);
-}
-
-static int	count_letters(char const *str, char c, int i)
-{
-	int	size;
-
-	size = 0;
-	while (str[i] != c && str[i])
-	{
-		size++;
-		i++;
-	}
-	return (size);
-}
-
-size_t	ft_strlen(const char *str)
-{
-	size_t	cont;
-
-	cont = 0;
-	while (str[cont] != '\0')
-	{
-		cont++;
-	}
-	return (cont);
-}
-
-char	*ft_substr(char const *s, unsigned int start, size_t len)
-{
-	char				*string;
-	unsigned int		i;
-	unsigned int		a;
-
-	i = start;
-	a = 0;
-	if (!s)
-		return (NULL);
-	if (start > ft_strlen(s))
-	{
-		string = malloc(1 * sizeof(char));
-		if (string == NULL)
-			return (NULL);
-		string[0] = '\0';
-		return (string);
-	}
-	string = malloc((len + 1) * sizeof(char));
-	if (string == NULL)
-		return (NULL);
-	while (i < ft_strlen(s) && len-- > 0)
-		string[a++] = s[i++];
-	string[a] = '\0';
-	return (string);
-}
-
-char	**ft_split(char const *s, char c)
-{
-	int		i;
-	int		j;
-	int		word;
-	char	**str;
-
-	if (!s)
-		return (NULL);
-	i = 0;
-	j = -1;
-	word = count_words(s, c);
-	str = (char **)malloc((word + 1) * sizeof(char *));
-	if (!str)
-		return (NULL);
-	while (++j < word)
-	{
-		while (s[i] == c)
-			i++;
-		str[j] = ft_substr(s, i, count_letters(s, c, i));
-		if (!str)
-			return (NULL);
-		i += count_letters(s, c, i);
-	}
-	str[j] = 0;
-	return (str);
-}
