--- srcs/general_operations.c
@@ -1,73 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   general_operations.c                               :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: pin3dev <pinedev@outlook.com>              +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/17 22:48:33 by ivbatist          #+#    #+#             */
-/*   Updated: 2024/08/08 15:05:43 by pin3dev          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../inc/push_swap.h"
-
-void	swap(t_stack **stack)
-{
-	t_stack	*tmp;
-
-	if (!*stack || !((*stack)->next))
-		return ;
-	tmp = *stack;
-	*stack = (*stack)->next;
-	tmp->next = (*stack)->next;
-	(*stack)->next = tmp;
-}
-
-void	rotate(t_stack **stack)
-{
-	t_stack	*tmp;
-
-	if (!*stack || !(*stack)->next)
-		return ;
-	tmp = *stack;
-	*stack = ft_lstlast(*stack);
-	(*stack)->next = tmp;
-	*stack = tmp->next;
-	tmp->next = NULL;
-}
-
-void	reverse_rotate(t_stack **stack)
-{
-	t_stack	*tmp;
-	int		i;
-
-	if (!*stack || !(*stack)->next)
-		return ;
-	i = 0;
-	tmp = *stack;
-	while ((*stack)->next)
-	{
-		*stack = (*stack)->next;
-		i++;
-	}
-	(*stack)->next = tmp;
-	while (i > 1)
-	{
-		tmp = tmp->next;
-		i--;
-	}
-	tmp->next = NULL;
-}
-
-void	push(t_stack **dest, t_stack **src)
-{
-	t_stack	*tmp;
-
-	if (!*src)
-		return ;
-	tmp = *dest;
-	*dest = *src;
-	*src = (*src)->next;
-	(*dest)->next = tmp;
-}
